Move tests for the file database clients to mirror where the files moved to in services.
